Move utils/count_oval_objects.py to ssg/_utils_count.py
