+++ ssg/__init__.py
@@ -20,6 +20,7 @@ import ssg._products as products
 import ssg._parse_oval as parse_oval
 import ssg._stig as stig
 import ssg._utils as utils
+import ssg._utils_count as utils_count
 import ssg._xccdf as xccdf
 import ssg._xml as xml
 import ssg._yaml as yaml
